# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pets:
    get:
      operationId: listPets
      tags: [pets]
      summary: List pets
      responses:
        "200":
          description: Pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"
    post:
      operationId: createPet
      tags: [pets]
      summary: Create a pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetInput"
      responses:
        "201":
          description: Created pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        "400":
          $ref: "#/components/responses/PetLimitReached"
  /pets/{petId}:
    get:
      operationId: getPet
      tags: [pets]
      summary: Get a pet
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "200":
          description: Pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updatePet
      tags: [pets]
      summary: Update a pet
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetUpdate"
      responses:
        "200":
          description: Updated pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
    delete:
      operationId: deletePet
      tags: [pets]
      summary: Delete a pet
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "204":
          description: Pet deleted
  /pets/{petId}/records:
    get:
      operationId: listHealthRecords
      tags: [pets]
      summary: List health records for a pet
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "200":
          description: Health records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HealthRecord"
    post:
      operationId: createHealthRecord
      tags: [pets]
      summary: Add a health record
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HealthRecordInput"
      responses:
        "201":
          description: Created health record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthRecord"
  /pets/{petId}/records/{recordId}:
    patch:
      operationId: updateHealthRecord
      tags: [pets]
      summary: Update a health record
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/HealthRecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HealthRecordUpdate"
      responses:
        "200":
          description: Updated health record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthRecord"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteHealthRecord
      tags: [pets]
      summary: Delete a health record
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/HealthRecordId"
      responses:
        "204":
          description: Health record deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /pets/{petId}/documents:
    post:
      operationId: uploadHealthRecordDocument
      tags: [pets]
      summary: Upload a health record document (PDF or image, 10MB max)
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "201":
          description: Uploaded document — path is a private storage key, not a usable URL by itself; attach it to a record via documentStoragePath, then fetch a viewable link on demand from GET .../document-url
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadResult"
        "400":
          description: File missing, too large, or an unsupported type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /pets/{petId}/records/{recordId}/document-url:
    get:
      operationId: getHealthRecordDocumentUrl
      tags: [pets]
      summary: Get a viewable link for a health record's attached document — a fresh short-lived signed URL for an uploaded file, or the stored link as-is for a pasted one
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/HealthRecordId"
      responses:
        "200":
          description: A link the client can open immediately; expires soon for uploaded documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUrlResponse"
        "404":
          description: Record not found, or it has no attached document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /pets/{petId}/medications:
    get:
      operationId: listMedications
      tags: [pets]
      summary: List medications for a pet
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "200":
          description: Medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [pets]
      summary: Add a medication
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationInput"
      responses:
        "201":
          description: Created medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /pets/{petId}/medications/{medicationId}:
    patch:
      operationId: updateMedication
      tags: [pets]
      summary: Update a medication
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/MedicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationUpdate"
      responses:
        "200":
          description: Updated medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          $ref: "#/components/responses/NotFound"
  /pets/{petId}/medications/{medicationId}/log-dose:
    post:
      operationId: logMedicationDose
      tags: [pets]
      summary: Mark a dose given and auto-calculate the next dose from the medication's interval
      parameters:
        - $ref: "#/components/parameters/PetId"
        - $ref: "#/components/parameters/MedicationId"
      responses:
        "200":
          description: Medication with updated next-dose time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "400":
          description: Medication has no structured interval to calculate from
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
  /pets/{petId}/reminders:
    get:
      operationId: listReminders
      tags: [pets]
      summary: List reminders for a pet
      parameters:
        - $ref: "#/components/parameters/PetId"
      responses:
        "200":
          description: Reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [pets]
      summary: Add a reminder
      parameters:
        - $ref: "#/components/parameters/PetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /reminders/{reminderId}/complete:
    post:
      operationId: completeReminder
      tags: [pets]
      summary: Complete a reminder
      parameters:
        - $ref: "#/components/parameters/ReminderId"
      responses:
        "200":
          description: Completed reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get the owner dashboard summary
      parameters:
        - $ref: "#/components/parameters/PetIdQuery"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
components:
  parameters:
    PetId:
      name: petId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PetLimitReached:
      description: Account has reached its pet limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Pet:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        species: { type: string, enum: [dog, cat, bird, rabbit, other] }
        breed: { type: ["string", "null"] }
        sex: { type: string, enum: [female, male, unknown] }
        birthDate: { type: ["string", "null"], format: date }
        weight: { type: ["number", "null"] }
        weightUnit: { type: string, enum: [lb, kg] }
        photoUrl: { type: ["string", "null"], format: uri }
        notes: { type: ["string", "null"] }
        vetName: { type: ["string", "null"] }
        vetClinic: { type: ["string", "null"] }
        vetPhone: { type: ["string", "null"] }
        vetAddress: { type: ["string", "null"] }
      required: [id, name, species, breed, sex, birthDate, weight, weightUnit, photoUrl, notes, vetName, vetClinic, vetPhone, vetAddress]
...
    PetUpdate:
      allOf:
        - $ref: "#/components/schemas/PetInput"
      description: Updated pet fields
...